+++ xmlschema/tests/test_validators.py
@@ -1430,7 +1430,7 @@ class TestEncoding(XsdValidatorTestCase):
 
         with self.assertRaises(XMLSchemaChildrenValidationError):
             schema.to_etree({"A": [1, 2], "B": [3, 4]})
-            
+
         root = schema.to_etree({"A": [1, 2], "B": [3, 4]}, converter=UnorderedConverter)
         self.assertListEqual([e.text for e in root], ['1', '3', '2', '4'])